feat: Excessões NotFound TicketManager

+++ ms-ticket-manager/src/main/java/io/github/juandev/msticketmanager/service/TicketService.java
@@ -1,6 +1,9 @@
 package io.github.juandev.msticketmanager.service;
 
+import feign.FeignException;
 import io.github.juandev.msticketmanager.client.EventClient;
+import io.github.juandev.msticketmanager.exception.BadRequestException;
+import io.github.juandev.msticketmanager.exception.NotFoundException;
 import io.github.juandev.msticketmanager.model.Event;
 import io.github.juandev.msticketmanager.model.Ticket;
 import io.github.juandev.msticketmanager.repository.TicketRepository;
@@ -8,6 +11,7 @@ import io.github.juandev.msticketmanager.web.dto.TicketDto;
 import io.github.juandev.msticketmanager.web.dto.mapper.TicketMapper;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +32,7 @@ public class TicketService {
 
     public Ticket findById(String ticketId) {
         return ticketRepository.findById(ticketId).orElseThrow(
-                () -> new RuntimeException("Ticket not found")
+                () -> new NotFoundException("Ticket not found")
         );
     }
 
@@ -36,17 +40,29 @@ public class TicketService {
     public Ticket save(TicketDto ticketDto) {
         Ticket ticket = ticketMapper.DtoToTicket(ticketDto);
 
-        Event event = eventClient.findById(ticketDto.getEventId());
-        ticket.setEvent(event);
-
-        return ticketRepository.save(ticket);
+        try{
+            Event event = eventClient.findById(ticketDto.getEventId());
+            ticket.setEvent(event);
+            return ticketRepository.save(ticket);
+        }
+        catch (FeignException.FeignClientException e){
+            throw new NotFoundException("Event not found");
+        }
+        //Depois ver a excessão dos bad requests que não está indo
     }
 
     public void deleteById(String ticketId) {
+        if (!ticketRepository.existsById(ticketId)) {
+            throw new NotFoundException("Ticket not found");
+        }
         ticketRepository.deleteById(ticketId);
     }
 
     public Ticket update(TicketDto ticketDto, String ticketId) {
+        if (!ticketRepository.existsById(ticketId)) {
+            throw new NotFoundException("Ticket not found");
+        }
+
         Ticket ticket = findById(ticketId);
         ticket.setCpf(ticketDto.getCpf());
         ticket.setCustomerName(ticketDto.getCustomerName());
@@ -58,10 +74,16 @@ public class TicketService {
         ticket.setEvent(event);
 
         return ticketRepository.save(ticket);
+
+        //Depois ver a excessão dos bad requests que não está indo
     }
 
     public List<Ticket> findByEvent(String eventId) {
-        Event event = eventClient.findById(eventId);
-        return ticketRepository.findByEvent(event);
+        try {
+            Event event = eventClient.findById(eventId);
+            return ticketRepository.findByEvent(event);
+        }catch (FeignException.FeignClientException e){
+            throw new NotFoundException("Event not found");
+        }
     }
 }
